PEP8: Renames _Tracker module
