+++ analysis.py
@@ -21,7 +21,7 @@ from moviepy.video.io.bindings import mplfig_to_npimage;
 from moviepy.editor import VideoFileClip, VideoClip, clips_array,ImageSequenceClip;
 
 import TopMouseTracker.utilities as utils;
-import TopMouseTracker._Tracker as tracker;
+import TopMouseTracker._tracker as tracker;
 
 #class GroupData() :
 #
